requirements.txt
imaplib
email
openai
smtplib
python-dotenv